requirements.txt
pandas
beautifulsoup4
selenium
webdriver-manager
google-cloud-bigquery
google-auth
